+++ estimator/__init__.py
@@ -1,0 +1,1 @@
+from estimator.appearance_model import estimate_gaze, model
